+++ index.html
@@ -117,7 +117,7 @@
                         <div class="saved-trips__cards">
 
                         </div>
-                        <div class="directions-btn">DIRECTIONS</div>
+                        <!-- <div class="directions-btn">DIRECTIONS</div> -->
                     </div>
                 </div>
             </div> <!-- closes "saved-trips"-->
